FIXED katon tail should find app regardless of directory
We’ve got a case where the project’s name (and so working directory)
doesn’t match the hostname.

`katon tail` can’t find the log file since it tries to use the working
directory name instead of the host name.

This change iterates over the hosts in ~/.katon to find the one that’s
associated with the current working directory.